python_ai/requirements.txt
twitchio>=2.6.0
openai>=1.0.0
python-dotenv>=1.0.0
elevenlabs>=1.0.0
websockets>=12.0
aiohttp>=3.9.0
gtts>=2.5.0

# Pokemon / Game AI (optional - install if using pokemon feature)
pyboy>=2.0.0
stable-baselines3>=2.0.0
torch>=2.0.0
transformers>=4.30.0
gymnasium>=0.29.0
numpy>=1.24.0
Pillow>=10.0.0

